refactor: move `getDefaultResolvedEnvironmentOptions` to `node/server/environment.ts`
breaks the following circular dependecy:
config.ts ->
server/environments/runnableEnvironment.ts ->
server/environment.ts ->
config.ts
